refactor: scholarship 도메인 구조 리팩토링
