--- src/components/ScholarshipApply/index.ts
@@ -1,6 +1,0 @@
-export { default as ApplySection } from './ApplySection';
-export { default as ApplySucceedModal } from './ApplySucceedModal';
-export { default as ConsentSection } from './ConsentSection';
-export { default as MileageBannerSection } from './MileageBannerSection';
-export { default as MileageCountBox } from './MileageCountBox';
-export { default as NotScholarshipDurationSection } from './NotScholarshipDurationSection';
