+++ src/hooks/queries/index.ts
@@ -1,13 +1,10 @@
 export { default as useDeleteSubmittedMileageMutation } from './useDeleteSubmittedMileageMutation';
 export { default as useGetEtcMileageQuery } from './useGetEtcMileageQuery';
-export { default as useGetIsAppliedScholarshipQuery } from './useGetIsAppliedScholarshipQuery';
-export { default as useGetScholarshipDurationQuery } from './useGetScholarshipDurationQuery';
 export { default as useGetSubmittedFileQuery } from './useGetSubmittedFileQuery';
 export { default as useGetSubmittedMileageQuery } from './useGetSubmittedMileageQuery';
 export { default as usePatchSubmittedMileageMutation } from './usePatchSubmittedMileageMutation';
 export { default as usePostLogoutMutation } from './usePostLogoutMutation';
 export { default as usePostNewMileageMutation } from './usePostNewMileageMutation';
-export { default as usePostScholarshipApplyMutation } from './usePostScholarshipApplyMutation';
 
 export * from './useGetUserInfoQuery';
 export * from './usePostLoginMutation';
