+++ src/pages/ScholarshipApplyPage/constants/personalInfoConsent.ts
