+++ src/pages/ScholarshipApplyPage/types/scholarship.ts
